requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]
certifi
scikit-learn
pyaml
dill
mlflow
dagshub

Flask
requests
fitbit
flask_sqlalchemy
requests_oauthlib
plotly
apache-airflow



## -e .